+++ Sources/VCore/Core/Utilities/Observation Containers/PublishingObservationContainer.swift
@@ -73,6 +73,10 @@ public final class PublishingObservationContainer<Value> {
     public init(_ value: Value) {
         self._value = value
     }
+
+    // Fixes compiler crash in `Swift` `6.3`
+    @_optimize(none)
+    deinit {}
 }
 
 #if DEBUG